+++ base/src/main/java/io/github/sds100/keymapper/base/utils/ui/compose/ComposeChipModel.kt
@@ -8,6 +8,10 @@ sealed class ComposeChipModel {
         override val id: String,
         val icon: ComposeIconInfo?,
         override val text: String,
+        /**
+         * Disabled chips are greyed out.
+         */
+        val isEnabled: Boolean = true,
     ) : ComposeChipModel()
 
     data class Error(